Rename page address
