+++ src/components/pages/android/install.tsx
